Spin Pound dive removal and Wario bug fix. (#11)
* Update character-movesets.lua

This new version removes the dive for Luigi's Spin Pound and fixes the momentum glitch with Wario

* Spin Pound Dive removal and Wario bug fix

The Spin Pound Dive was easily removed by just removing the code that puts you into a dive when B is pressed. The Wario bug was also fixed by simply making it so if Wario is in the Water Jump Hold action the faster air speed from holding an item doesn't apply to him.