Fix incorrect value for Makefile option (#269)
The comment for TEXTURE_FIX states "Disable texture fixes by default (helps with them purists)" and in spite of that, TEXTURE_FIX is set to 1 (on)
This PR simply sets it back to 0 as the comment intends the value to be, I just noticed this and saw an opportunity to make a PR for it. This doesn't have to be merged and all it does is restore the old smoke and pink on the Koopa's back.