Reorganize Info panel and Misc. (#14)
Makes the info credits paginated so we can add more text over time with no issues and also moves the Open Appdata/User Folder button to the Misc. menu to avoid cluttering.

Also fixed the size of the changelog panel as the Back button was also being obscured.

Rewritten it from the old repo, so all new changes should stay there.

There IS an issue where if you go to the changelog, and then back to info and try to go to another page the game crashes, and I think it's DJUI and not panel related, but I'm still looking into what it could be.