Boy what the hell (unbreak mods and DynOS packs)
They dug me out of my grave and told me `configSkipPackGeneration`. My heart sank.
This feature would have been fine if it was off by default and if it wasn't then people caught it before the version was officially released and the option was promptly disabled without a massive pool of people having this option turned on by default.
So I will say, Coolio wasn't really in the wrong by adding this, it should have never been changed to be on by default.