Add get_lua_function_id(stack_depth)
Returns a number that uniquely identifies a lua function that called
get_lua_function_id(stack_depth), or returns 0 on failure.

A stack_depth of 0 returns an identifer for the function that called
get_lua_function_id(0).

A stack_depth of 1 returns an identifier for the function that called
the function that called `get_lua_function_id(1)`. etc.