Prevent cursor interpolation from causing crashes
Previously djui_cursor_interp() called djui_cursor_update_position(), which
is really bad, because the cursor can trigger events that would affect the
display lists. The number of things in the display list must *not* change
during interpolation, or we corrupt the list.

Steps to cause the crash before this change are simple, but maybe only on
windows:
1. Open game
2. Open console
3. Wiggle mouse on console