Fix crash on input box (#1043)
When typing ` into the coopnet password inputbox (the same as the console bind)
It would do all of the null checks, then open the console changing the focus,
then call a null function because the null checks are no longer valid.

So I moved the null checks to where they're needed.

Co-authored-by: MysterD <myster@d>