Update djui_chat_message.c
If we replace MAX_CHAT_PACKET_LENGTH in these two spots, it fixes the previous issue of having the chat text entry spill over the character limit (resulting in characters at the end of messages that are lost) while raising the character limit to the intended limit of 500. Or 499 more specifically.